Cleanup imports & move types together
