+++ packages/lib-utils/src/webSocket/WSFactory.ts
@@ -1,73 +1,17 @@
 import { consoleLogger } from '@monorepo/common';
 import type {
+  BulkMessageHandler,
+  CloseHandler,
+  DestroyHandler,
   EventHandlers,
   EventHandlerTypes,
-  CloseHandler,
-  MessageHandler,
-  BulkMessageHandler,
-  MessageDataType,
-  DestroyHandler,
-  OpenHandler,
   ErrorHandler,
+  MessageDataType,
+  MessageHandler,
+  OpenHandler,
+  WSOptions,
 } from './types';
 import { applyConfigSubProtocols, applyConfigHost, createURL } from './ws-utils';
-
-/**
- * The web socket configuration options.
- */
-export type WSOptions = {
-  /**
-   * The path to the resource you wish to watch.
-   */
-  path: string;
-
-  /**
-   * Overridable web socket host URL for plugins. Normally set by the application.
-   */
-  host?: string;
-
-  /**
-   * Overridable web socket sub protocols for plugins. Normally set by the application.
-   * Note: This is ignored if `host` is not set.
-   */
-  subProtocols?: string[];
-
-  /**
-   * Set to true if you want automatic reconnection if it fails to create or when the web socket
-   * closes.
-   */
-  reconnect?: boolean;
-
-  /**
-   * Set to true if you wish to get your data back in JSON format when the WS sends a message.
-   * Note: If it's not valid JSON, a warning will be logged and you get back the raw message.
-   */
-  jsonParse?: boolean;
-
-  /**
-   * Set a maximum buffer to hold onto between `bufferFlushInterval`s. Messages that exceed the
-   * buffer are dropped.
-   *
-   * Unit is in number of messages.
-   */
-  bufferMax?: number;
-
-  /**
-   * Configure a duration between messages being flushed out in events.
-   *
-   * Note: If `bufferMax` is not set, this is ignored.
-   * Defaults to 500ms.
-   */
-  bufferFlushInterval?: number;
-
-  /**
-   * Set a connection limit for when to give up on the current instance of the web socket.
-   *
-   * If omitted, the web socket will continue to try to reconnect only if you set the `reconnect`
-   * flag.
-   */
-  timeout?: number;
-};
 
 /**
  * States the web socket can be in.
